Use long for loop exit in SWI kernels to avoid possible integer overflow

+++ fpga-stream-kernel-blk2d.cl
@@ -48,9 +48,9 @@ __kernel void mac(__global const float* restrict a, __global const float* restri
 #else // Single Work-item kernels
 
 __attribute__((max_global_work_dim(0)))
-__kernel void copy(__global const float* restrict a, __global float* restrict c, const int pad, const int dim_y, const int dim_x, const int exit, const int halo)
+__kernel void copy(__global const float* restrict a, __global float* restrict c, const int pad, const int dim_y, const int dim_x, const long exit, const int halo)
 {
-	int cond = 0;
+	long cond = 0;
 	int x = 0;
 	int y = 0;
 	int bx = 0;
@@ -89,9 +89,9 @@ __kernel void copy(__global const float* restrict a, __global float* restrict c,
 }
 
 __attribute__((max_global_work_dim(0)))
-__kernel void mac(__global const float* restrict a, __global const float* restrict b, __global float* restrict c, const float constValue, const int pad, const int dim_y, const int dim_x, const int exit, const int halo)
+__kernel void mac(__global const float* restrict a, __global const float* restrict b, __global float* restrict c, const float constValue, const int pad, const int dim_y, const int dim_x, const long exit, const int halo)
 {
-	int cond = 0;
+	long cond = 0;
 	int x = 0;
 	int y = 0;
 	int bx = 0;
